Add instructions to import a new VistA-FOIA from CACHE.DAT
Write instructions to cover the mechanics of extracting routines and
globals from CACHE.DAT, unpacking them in the source tree, populating
the package directory structure, and committing to Git.  Place the
instructions in the repository next to the corresponding scripts so they
can be updated together.  For now leave out details of managing a
dedicated Git branch to contain unmodified VistA FOIA versions and avoid
overwriting changes that have been made in the OSEHRA version.

Currently we do not handle the Cache-specific xobw.* classes from the
Web Services Client or their intermediate Cache ObjectScript routines.
The routines were left out of the initial import of VistA-FOIA.  For now
leave a TODO note along with instructions to remove the routines.

Change-Id: I6fc9cb6a4759e7f90c3b7a3ed90e705071fb05ea